career_app_backend/requirements.txt
# Flask Framework and Extensions
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# Environment and Configuration
python-dotenv==1.0.0

# OpenAI for AI Analysis (used directly, not via AutoGen)
openai

# PDF Processing - Cloud-friendly alternatives
PyPDF2==3.0.1
pdfplumber==0.9.0

# File Handling and Utilities
werkzeug==2.3.7
requests==2.31.0

# Payment Processing
razorpay==1.4.1

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0
